Fix cycle import.
